Fix tests broken by upcoming Test::More
circular::require needs to be loaded before base/parent/etc. However
Test::More will soon load base. This makes sure circular::require is the
first thing loaded, even before strict/warnings.

This is a quick fix to get the tests passing again. Ideally the code
would check to see what is already loaded, and warn for some obvious
ones like base/parent.

It would also be good if the code detected PERL5OPT and similar that can
throw this off. It really needs to be good about ensuring it is the
first thing loaded.